Reorganice files, added styles to profile page
