--- src/components/DragonProfile/DragonProfile.module.scss
@@ -1,1 +1,0 @@
-
