+++ src/components/Mission/MissionPage.js
@@ -3,7 +3,6 @@ import { useSelector, useDispatch } from 'react-redux';
 import { fetchMissions, reservationUpdates } from '../../Redux/mission';
 import Navigation from '../Navigation/Navigation';
 import './missionPage.css';
-// import { reservationUpdates } from '../../Redux/mission';
 
 const Missions = () => {
   const getMissions = useSelector((state) => state.missions);
