--- src/components/Mission/ProfilePage.js
@@ -1,47 +1,0 @@
-import React from 'react';
-import { useSelector } from 'react-redux';
-import Navigation from '../Navigation/Navigation';
-import './profile.css';
-import RocketStatus from '../Profile/RocketStatus';
-import DragonProfile from '../DragonProfile/DragonProfile';
-
-function ProfilePage() {
-  const missionProfile = useSelector((state) => state.missions);
-  const missions = missionProfile;
-  const missionsReserved = missions.filter((data) => data.reserved === true);
-  let missionCounter = 0;
-
-  missions.map((data) => {
-    if (data.reserved) {
-      missionCounter += 1;
-    }
-    return missionCounter;
-  });
-  return (
-    <div>
-      <Navigation />
-      <main className="container">
-        <section className="profile-sect">
-          <h1 className="header">My Missions</h1>
-          <ul className="missionsBooked">
-            {missionCounter > 0 ? (
-              missionsReserved.map((data) => (
-                <li className="missionLi" key={data.mission_id}>
-                  {data.mission_name}
-                </li>
-              ))
-            ) : (
-              <li className="hide" style={{ border: 'none' }}>
-                reserved missions currently(0)
-              </li>
-            )}
-          </ul>
-        </section>
-        <RocketStatus />
-        <DragonProfile />
-      </main>
-    </div>
-  );
-}
-
-export default ProfilePage;
