+++ src/components/Profile/profile.css
@@ -3,11 +3,27 @@
 .container {
   display: flex;
   align-items: center;
-  justify-content: center;
   font-family: 'Roboto', sans-serif;
-  column-gap: 2rem;
+  column-gap: 1rem;
+  padding: 1rem 2rem;
 }
 
-.container ul {
+.profile-sect {
+  flex-basis: 33.3%;
+}
+
+.profile-sect h1 {
+  font-size: 1.5rem;
+  margin-bottom: 0.7rem;
+  text-align: center;
+}
+
+.container ul{
   list-style: none;
+}
+
+.container ul li {
+  border: 1px solid grey;
+  padding: 1rem 0;
+  text-align: center;
 }
